Optimised object mapping. Removed 'publish' option. Added yarn. Added IViewResolver

+++ src/Kasbah/Content/Models/Node.cs
@@ -10,7 +10,7 @@ namespace Kasbah.Content.Models
         public string Alias { get; set; }
         public string DisplayName { get; set; }
         public string Type { get; set; }
-        public int? PublishedVersion { get; set; }
+        public long? PublishedVersion { get; set; }
         public DateTime Created { get; set; }
         public DateTime Modified { get; set; }
     }
